Created fragment for battery service
