+++ app/src/main/java/io/github/webbluetoothcg/bletestperipheral/BatteryServiceFragment.java
@@ -1,0 +1,133 @@
+package io.github.webbluetoothcg.bletestperipheral;
+
+import android.app.Fragment;
+import android.bluetooth.BluetoothGattCharacteristic;
+import android.bluetooth.BluetoothGattService;
+import android.os.Bundle;
+import android.view.KeyEvent;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.view.inputmethod.EditorInfo;
+import android.widget.EditText;
+import android.widget.SeekBar;
+import android.widget.SeekBar.OnSeekBarChangeListener;
+import android.widget.TextView;
+import android.widget.TextView.OnEditorActionListener;
+import android.widget.Toast;
+
+import java.util.UUID;
+
+
+public class BatteryServiceFragment extends Fragment {
+
+  ///////////////////////
+  ////// Constants //////
+  ///////////////////////
+  private static final UUID BATTERY_SERVICE_UUID = UUID
+      .fromString("0000180F-0000-1000-8000-00805f9b34fb");
+
+  private static final UUID BATTERY_LEVEL_UUID = UUID
+      .fromString("00002A19-0000-1000-8000-00805f9b34fb");
+  private static final int INITIAL_BATTERY_LEVEL = 50;
+  private static final int BATTERY_LEVEL_MAX = 100;
+
+  ////////////////
+  ////// UI //////
+  ////////////////
+  private EditText mBatteryLevelEditText;
+  private final OnEditorActionListener mOnEditorActionListener = new OnEditorActionListener() {
+    @Override
+    public boolean onEditorAction(TextView textView, int actionId, KeyEvent event) {
+      if (actionId == EditorInfo.IME_ACTION_DONE) {
+        String newBatteryLevelString = textView.getText().toString();
+        // Need to check if the string is empty since isDigitsOnly returns
+        // true for empty strings.
+        if (!newBatteryLevelString.isEmpty()
+            && android.text.TextUtils.isDigitsOnly(newBatteryLevelString)) {
+          int newBatteryLevel = Integer.parseInt(newBatteryLevelString);
+          if (newBatteryLevel <= BATTERY_LEVEL_MAX) {
+            setBatteryLevel(newBatteryLevel, textView);
+          } else {
+            Toast.makeText(getActivity(), R.string.batteryLevelTooHigh, Toast.LENGTH_SHORT)
+                .show();
+          }
+        } else {
+          Toast.makeText(getActivity(), R.string.batteryLevelIncorrect, Toast.LENGTH_SHORT)
+              .show();
+        }
+      }
+      return false;
+    }
+  };
+  private SeekBar mBatteryLevelSeekBar;
+  private final OnSeekBarChangeListener mOnSeekBarChangeListener = new OnSeekBarChangeListener() {
+    @Override
+    public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
+      if (fromUser) {
+        setBatteryLevel(progress, seekBar);
+      }
+    }
+
+    @Override
+    public void onStartTrackingTouch(SeekBar seekBar) {
+
+    }
+
+    @Override
+    public void onStopTrackingTouch(SeekBar seekBar) {
+
+    }
+  };
+
+  //////////////////
+  ////// GATT //////
+  //////////////////
+  private BluetoothGattService mBatteryService;
+  private BluetoothGattCharacteristic mBatteryLevelCharacteristic;
+
+  public BatteryServiceFragment() {
+    mBatteryLevelCharacteristic =
+        new BluetoothGattCharacteristic(BATTERY_LEVEL_UUID,
+            BluetoothGattCharacteristic.PROPERTY_READ | BluetoothGattCharacteristic.PROPERTY_NOTIFY,
+            BluetoothGattCharacteristic.PERMISSION_READ);
+
+    mBatteryService = new BluetoothGattService(BATTERY_SERVICE_UUID,
+        BluetoothGattService.SERVICE_TYPE_PRIMARY);
+    mBatteryService.addCharacteristic(mBatteryLevelCharacteristic);
+  }
+
+  /////////////////////////////////
+  ////// Lifecycle callback //////
+  /////////////////////////////////
+  @Override
+  public View onCreateView(LayoutInflater inflater, ViewGroup container,
+      Bundle savedInstanceState) {
+
+    View view = inflater.inflate(R.layout.fragment_battery, container, false);
+
+    mBatteryLevelEditText = (EditText) view.findViewById(R.id.textView_batteryLevel);
+    mBatteryLevelEditText.setOnEditorActionListener(mOnEditorActionListener);
+    mBatteryLevelSeekBar = (SeekBar) view.findViewById(R.id.seekBar_batteryLevel);
+    mBatteryLevelSeekBar.setOnSeekBarChangeListener(mOnSeekBarChangeListener);
+
+    setBatteryLevel(INITIAL_BATTERY_LEVEL, null);
+    return view;
+  }
+
+  public BluetoothGattService getBluetoothGattService() {
+    return mBatteryService;
+  }
+
+  private void setBatteryLevel(int newBatteryLevel, View source) {
+    mBatteryLevelCharacteristic.setValue(newBatteryLevel,
+        BluetoothGattCharacteristic.FORMAT_UINT8, /* offset */ 0);
+    if (source != mBatteryLevelSeekBar) {
+      mBatteryLevelSeekBar.setProgress(newBatteryLevel);
+    }
+    if (source != mBatteryLevelEditText) {
+      mBatteryLevelEditText.setText(Integer.toString(newBatteryLevel));
+    }
+  }
+
+}
