+++ app/src/main/res/layout/activity_peripherals.xml
@@ -41,39 +41,10 @@
             android:layout_height="wrap_content"
             android:layout_below="@id/label_connectionStatus"/>
 
-    <TextView
-            android:id="@+id/textView_serviceName"
-            android:text="@string/batteryService_name"
-            android:textAppearance="@style/subheader"
-            android:gravity="center_vertical"
-            android:layout_width="wrap_content"
-            android:layout_height="48dp"
-            android:layout_below="@id/textView_connectionStatus"/>
-    <LinearLayout
-            android:id="@+id/batteryLevel"
-            android:orientation="horizontal"
-            android:gravity="center_vertical"
+    <FrameLayout
+            android:id="@+id/fragment_container"
             android:layout_width="fill_parent"
             android:layout_height="wrap_content"
-            android:layout_below="@id/textView_serviceName">
-        <ImageView
-                android:id="@+id/imageView_batteryIcon"
-                android:src="@drawable/battery_icon"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"/>
-        <SeekBar
-                android:id="@+id/seekBar_batteryLevel"
-                android:layout_weight="1"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"/>
-        <EditText
-                android:id="@+id/textView_batteryLevel"
-                android:ems="3"
-                android:maxLength="3"
-                android:gravity="center"
-                android:inputType="number"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"/>
-    </LinearLayout>
+            android:layout_below="@id/textView_connectionStatus"/>
 
 </RelativeLayout>
