+++ app/src/main/res/layout/fragment_battery.xml
@@ -1,0 +1,42 @@
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+             xmlns:tools="http://schemas.android.com/tools"
+             android:layout_width="match_parent"
+             android:layout_height="match_parent"
+             tools:context="io.github.webbluetoothcg.bletestperipheral.BatteryServiceFragment">
+    <RelativeLayout android:layout_width="fill_parent"
+                    android:layout_height="wrap_content">
+        <TextView
+                android:id="@+id/textView_serviceName"
+                android:text="@string/batteryService_name"
+                android:textAppearance="@style/subheader"
+                android:gravity="center_vertical"
+                android:layout_width="wrap_content"
+                android:layout_height="48dp"/>
+        <LinearLayout
+                android:id="@+id/batteryLevel"
+                android:orientation="horizontal"
+                android:gravity="center_vertical"
+                android:layout_width="fill_parent"
+                android:layout_height="wrap_content"
+                android:layout_below="@id/textView_serviceName">
+            <ImageView
+                    android:id="@+id/imageView_batteryIcon"
+                    android:src="@drawable/battery_icon"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"/>
+            <SeekBar
+                    android:id="@+id/seekBar_batteryLevel"
+                    android:layout_weight="1"
+                    android:layout_width="0dp"
+                    android:layout_height="wrap_content"/>
+            <EditText
+                    android:id="@+id/textView_batteryLevel"
+                    android:ems="3"
+                    android:maxLength="3"
+                    android:gravity="center"
+                    android:inputType="number"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"/>
+        </LinearLayout>
+    </RelativeLayout>
+</FrameLayout>
