+++ app/src/main/res/values/strings.xml
@@ -24,4 +24,5 @@
     <string name="batteryService_name">Battery Service</string>
     <string name="batteryLevelTooHigh">Batter Level shouldn\'t exceed 100</string>
     <string name="batteryLevelIncorrect">Please enter a number between 0 and 100</string>
+
 </resources>
